refactor: separete config cmd and basic cmd
